Add files via upload
This system provides an intuitive and interactive GUI through which various functions can be accessed. The frontend of the system which is the GUI, is coded in Visual C++ on the Microsoft .NET framework in the Visual Studio IDE. The backend of the system for image analysis and processing is implemented in OpenCV C++. The system supports various digital image formats such as JPG, BMP, GIF, PNG and many others. Moreover, this system enables the common functions, such as noise reduction, image deblurring, image enhancement, edge detection, as well as various other image manipulation/transformation operations. In addition, this system accomplishes special image effects, including image grayscaling, kaleidoscope, Cartooning, erosion, etc